agent_backend/requirements.txt
openai-agents>=0.0.14
fastapi>=0.115.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0 